pyproject.toml
[tool.poetry]
name = "web-scraping"
version = "0.1.0"
description = ""
authors = ["Rômulo do Carmo Sousa <carmosousa20@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-dotenv = "^0.15.0"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
